added 2 folders and 2 files and added code to the index.html & main.css files
created Mod2_solution's folder with a css folder inside of that.

created index.html file inside of Mod2_solution folder and main.css file inside of Mod2_solution\css folder path.

created a title for the index.html file and a h1 tag for the heading

created a div with a class of "row" and inside that div I created 3 more div tags with classes for the columns

created a span tag inside of the div tag for the name of each div block.

created a 'p' tag and added some filler text (Lorem ipsum) inside of the div tags.

added in styles to separate each of the different div blocks and adjusted the heading of the page to be centered.